+++ UnitTest_PZ/TestRefinement/TestRefinement.cpp
@@ -110,7 +110,7 @@ namespace refinementtests{
             if(!gel) continue;
             if(gel->HasSubElement()) continue;
 
-            TPZVec<REAL> qsi(gel->Dimension(),0.); TPZFMatrix<double> jac; TPZFMatrix<double> axes; REAL detjac = notcomputed; TPZFMatrix<double> jacinv;
+            TPZVec<REAL> qsi(gel->Dimension(),0.); TPZFMatrix<REAL> jac; TPZFMatrix<REAL> axes; REAL detjac = notcomputed; TPZFMatrix<REAL> jacinv;
             gel->Jacobian(qsi,jac,axes,detjac,jacinv);
             REAL zero = ZeroTolerance();
             bool cond = detjac > zero;
